requirements.txt
Django==2.2.19
nose==1.3.7
six==1.15.0
sqlparse==0.3.1
psycopg2-binary==2.8.6
djangorestframework==3.12.2
markdown==3.3.4
django-filter==2.4.0
djangorestframework-api-key==2.0.0
uWSGI>=2.0.18,<2.1